--- src/main/java/com/nhs2304/demosortalgo/HistoryEntry.java
@@ -1,10 +1,0 @@
-package com.nhs2304.demosortalgo;
-
-public class HistoryEntry {
-    String algorithm;
-    String inputArray;
-    String sortedArray;
-    long durationMillis;
-    int swapCount;
-    String timestamp;
-}
